requirements.txt
boto3==1.28.38
urllib3==1.26.16
requests==2.31.0
Pillow==10.0.0
PyMuPDF==1.22.5